projectile

+++ Assets/Script/SpawnPlayer.cs
@@ -22,47 +22,49 @@ public class SpawnPlayer : MonoBehaviourPunCallbacks
     // Start is called before the first frame update
     void Start()
     {
-        playerCount = PhotonNetwork.CurrentRoom.PlayerCount;
-        teamPosition = PlayerPrefs.GetInt("previousIndex");
+        if (PhotonNetwork.IsConnected) {
+            playerCount = PhotonNetwork.CurrentRoom.PlayerCount;
+            teamPosition = PlayerPrefs.GetInt("previousIndex");
 
-        PlayerPrefs.DeleteKey("previousIndex");
+            PlayerPrefs.DeleteKey("previousIndex");
 
-        Debug.Log("Number of player: " + playerCount);
-        Debug.Log("Position Selected: " + teamPosition);
+            Debug.Log("Number of player: " + playerCount);
+            Debug.Log("Position Selected: " + teamPosition);
 
-        for(int i = 0; i< playerPos.Length; i++)
-        {
-            if(teamPosition == i)
+            for(int i = 0; i< playerPos.Length; i++)
             {
-                playerName[i] = PhotonNetwork.NickName;
-                photonView.RPC("Set_OtherPlayerName", RpcTarget.OthersBuffered, i, PhotonNetwork.NickName);
-
-                Quaternion rotationB;
-                string color;
-                Color nameColor;
-
-                if (i > 2)
+                if(teamPosition == i)
                 {
-                   rotationB = Quaternion.Euler(0f, 180f, 0f);
-                    nameColor = Color.red;
-                    color = "Red";
+                    playerName[i] = PhotonNetwork.NickName;
+                    photonView.RPC("Set_OtherPlayerName", RpcTarget.OthersBuffered, i, PhotonNetwork.NickName);
+
+                    Quaternion rotationB;
+                    string color;
+                    Color nameColor;
+
+                    if (i > 2)
+                    {
+                    rotationB = Quaternion.Euler(0f, 180f, 0f);
+                        nameColor = Color.red;
+                        color = "Red";
+                    }
+                    else
+                    {
+                    rotationB =  Quaternion.identity;
+                        nameColor = Color.blue;
+                        color = "Blue";
+                    }
+
+                    GameObject spawnPlayer = PhotonNetwork.Instantiate(player.name, playerPos[i], rotationB);
+
+                    TextMeshProUGUI playername = spawnPlayer.GetComponentInChildren<TextMeshProUGUI>();
+                    playername.text = playerName[i];
+                    playername.color = nameColor;
+
+                    photonView.RPC("UpdatePlayerName", RpcTarget.OthersBuffered, playerName[i], color);
+
+                    break;
                 }
-                else
-                {
-                   rotationB =  Quaternion.identity;
-                    nameColor = Color.blue;
-                    color = "Blue";
-                }
-
-                GameObject spawnPlayer = PhotonNetwork.Instantiate(player.name, playerPos[i], rotationB);
-
-                TextMeshProUGUI playername = spawnPlayer.GetComponentInChildren<TextMeshProUGUI>();
-                playername.text = playerName[i];
-                playername.color = nameColor;
-
-                photonView.RPC("UpdatePlayerName", RpcTarget.OthersBuffered, playerName[i], color);
-
-                break;
             }
         }
     }
